Fix unittests
They have actually found two minor bugs:

1) When scaling down with NN filter there should be at least one
   pixel left in the destination space
2) KisTransformProcessingVisitor should also support fractional
   offsets, since it is supported by the worker now.

+++ libs/image/processing/kis_transform_processing_visitor.cpp
@@ -43,7 +43,7 @@ KisTransformProcessingVisitor(qreal  xscale, qreal  yscale,
                               qreal  xshear, qreal  yshear,
                               const QPointF &shearOrigin,
                               qreal angle,
-                              qint32  tx, qint32  ty,
+                              qreal  tx, qreal  ty,
                               KisFilterStrategy *filter,
                               const QTransform &shapesCorrection)
     : m_sx(xscale), m_sy(yscale)
